+++ CHANGELOG.md
@@ -41,6 +41,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add injury dice and enhancement for Near Death level
 - Add integrity rolls to Accursed
 - Increase the Defense and Integrity values of Accursed when buying Dodge/Resist tracks for their respective rolls
+- Increase and decrease Momentum value when purchasing or adjusting the Bolster trick
 
 ### Changed